+++ packages/web-client/vite.config.ts
@@ -3,7 +3,7 @@ import react from '@vitejs/plugin-react';
 
 export default defineConfig({
   plugins: [react()],
-  base: '/owa/',
+  base: '/',
   server: {
     port: 4000,
     proxy: {